Add support for `click_listener` in `GsiButtonConfiguration`
Google has added a new `click_listener` option to the `GsiButtonConfiguration` that allows users to specify a function to be called when the Sign in with Google button is called. Adding support for this property enables users of the library to recognize the start of an authentication flow, based on which other components' state may be reset.

Refs https://developers.google.com/identity/gsi/web/reference/js-reference#click_listener

Closes https://github.com/MomenSherif/react-oauth/issues/40