fix: fix error produced when passing width size to GoogleLogin
When user passed width to GoogleLogin, they might forget to put the px, it throws an error message
to the client but the msg was unclear. I tried to put number as well but typescript gave warning
that type 'number' can't be put to 'string'. This PR handled the error by asking user to add px to
the width if they wanted to put it as string and handle the 'number' is not 'string' issue by
converting the number to string px.